webapp/requirements.txt
fastapi==0.115.6
uvicorn[standard]==0.32.1
python-multipart==0.0.20
tensorflow==2.19.0
numpy==1.26.4
pillow==11.1.0


